migrate websocket connection state update
