+++ packages/wallet-sdk/src/relay/WalletLinkRelay.ts
@@ -123,8 +123,8 @@ export class WalletLinkRelay
 
     const { linkAPIUrl, diagnostic } = this;
     const connection = new WalletLinkConnection({
-      session,
       linkAPIUrl,
+      session,
       diagnostic,
       listener: this,
     });
@@ -132,9 +132,9 @@ export class WalletLinkRelay
     const { version, darkMode } = this.options;
     const ui = this.options.uiConstructor({
       linkAPIUrl,
+      session,
       version,
       darkMode,
-      session,
     });
 
     connection.connect();
